Use `parcel` to build (#11)
https://pagespeed.web.dev/analysis/https-quentinms-github-io-numeros/ly9hhr2lwc?form_factor=mobile
complains that we make too many requests.

Use [parcel](https://parceljs.org) to bundle + minify both JS and CSS.
It also updates the index.html to point to the new JS and CSS files, and
puts everything in `dist/`.
We need a dummy index.html at the root to redirect to the one in `dist/`
(this is a GitHub Pages limitation).